# architecture/api_specification.yaml
openapi: 3.0.0
info:
  title: Stablecoin Escrow API
  version: 1.0.0
  description: |
    API for managing privacy-preserving stablecoin escrows on DAML with Oracle integration.
    This version implements a schema-driven metadata model where core escrow dynamics
    are generic, and industry-specific contracted elements are validated via external JSON Schemas.

# 1. Top-level security requirement: All endpoints require bearerAuth by default.
security:
  - bearerAuth: []

paths:
  # 2. Auth-related paths
  /api/v1/auth/discover:
    post:
      summary: Discover Identity Provider
      description: |
        Performs Home Realm Discovery (HRD) based on the user's email domain
        to determine the correct OIDC/SAML issuer for authentication.
      operationId: discoverProvider
      tags:
        - auth
      # This endpoint is public.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryRequest'
      responses:
        '200':
          description: Authentication provider details retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProvider'
        '400':
          description: Invalid email format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /escrows:
    post:
      summary: Create a new escrow contract
      operationId: createEscrow
      tags:
        - escrows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEscrowRequest'
      responses:
        '201':
          description: Escrow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escrow'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /escrows/{escrowID}:
    get:
      summary: Get escrow details by ID
      operationId: getEscrow
      tags:
        - escrows
      parameters:
        - name: escrowID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Escrow details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escrow'
        '404':
          description: Escrow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /webhooks/milestone:
    post:
      summary: Oracle trigger for milestone approval
      description: |
        Receives external events to automatically advance escrow milestones.
        Payloads include structured evidence and can be verified against the
        contract's specific JSON Schema.
      operationId: oracleMilestoneTrigger
      tags:
        - oracles
      # This endpoint is public, but secured via HMAC signature check.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OracleWebhookRequest'
      responses:
        '200':
          description: Webhook processed and milestone approved
        '401':
          description: Invalid signature
        '404':
          description: Escrow or Milestone not found

components:
  # 3. Security Scheme definition
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A JWT issued by the configured OIDC provider (e.g., Okta).

  schemas:
    # Auth Schemas
    DiscoveryRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "user@gmail.com"
    
    AuthProvider:
      type: object
      properties:
        type:
          type: string
          enum: [OIDC, SAML]
          example: "OIDC"
        issuer:
          type: string
          format: uri
          example: "https://<your-okta-domain>/oauth2/default"
        tenantId:
          type: string
          description: The tenant ID for multi-tenant SAML/OIDC configurations.
        loginUrl:
          type: string
          format: uri
          description: The specific URL to redirect the user to for login.
        description:
          type: string
          example: "Okta Identity Provider"

    # Core Schemas
    CreateEscrowRequest:
      type: object
      required:
        - issuer
        - mediator
        - totalAmount
        - currency
        - milestones
      properties:
        issuer:
          type: string
        buyer:
          type: string
          description: Optional for anonymous buyer scenarios
        seller:
          type: string
          description: Optional for anonymous seller scenarios
        mediator:
          type: string
        totalAmount:
          type: number
          format: decimal
        currency:
          type: string
        description:
          type: string
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/Milestone'
        metadata:
          $ref: '#/components/schemas/EscrowMetadata'

    Escrow:
      type: object
      properties:
        id:
          type: string
        issuer:
          type: string
        buyer:
          type: string
        seller:
          type: string
        mediator:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        description:
          type: string
        currentMilestoneIndex:
          type: integer
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/Milestone'
        metadata:
          $ref: '#/components/schemas/EscrowMetadata'

    Milestone:
      type: object
      required:
        - label
        - amount
      properties:
        label:
          type: string
        amount:
          type: number
        completed:
          type: boolean
          default: false

    OracleWebhookRequest:
      type: object
      required:
        - escrowId
        - milestoneIndex
        - event
        - signature
      properties:
        escrowId:
          type: string
        milestoneIndex:
          type: integer
        event:
          type: string
        oracleProvider:
          type: string
        evidence:
          type: string
        metadata:
          type: object
          description: Raw event metadata payload
          additionalProperties: true
        signature:
          type: string

    EscrowMetadata:
      type: object
      description: |
        A schema-driven data container. The 'payload' is an 'any:any' map
        that should be marshalled according to the provided 'schemaUrl'.
      properties:
        schemaUrl:
          type: string
          description: URL to the JSON Schema for the payload (e.g., leasing.json, supply-chain.json)
          example: "https://stablecoin-escrow.io/schemas/leasing.v1.json"
        payload:
          type: object
          description: The actual metadata fields (Contracted Elements)
          additionalProperties: true
          example:
            assetId: "SERIAL-123"
            securityDepositAmount: 5000
            assetType: "Machinery"
    
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
          format: int32